ref: tighten remote acquisition ownership
