+++ src-tauri/src/output_artifact/mod.rs
@@ -20,7 +20,8 @@ pub(crate) use parent_dirs::{ensure_output_parent_dirs, OutputParentDirCleanup};
 pub(crate) use plan::OutputPlanLedger;
 pub(crate) use review::{enforce_output_plan_review, OutputPlanReview};
 pub(crate) use supplemental::{
-    commit_supplemental_output_asset, SupplementalOutputAssetCommitRequest,
+    commit_supplemental_output_asset, commit_supplemental_output_assets_for_output,
+    SupplementalOutputAssetCommitRequest, SupplementalOutputAssetsCommitRequest,
 };
 pub(crate) use types::ResolvedOutputPlan;
 pub use types::{
